Add files via upload
Objective of this assignment was given a x,y grid with colored pixels, check if it is feasible to cover all the colored pixels using a 2x1 pixel blocks. These blocks would not be allowed to be placed over noncolored pixels. It would return impossible if it is impossible to do so. If it is possible, it will provide the x,y coordinates of each 2x1 block that would be used to cover the grid.